Safari-611.1.10/LayoutTests/js/dom/weakmap-gc-unobservable-dom-nodes-expected.txt

PASS 4 is defined.
PASS 4 is defined.
PASS 4 is defined.
PASS 4 is defined.
PASS 4 is defined.
PASS successfullyParsed is true

TEST COMPLETE
